requirements.txt
cvxopt==1.3.2
matlab==0.1
matplotlib==3.9.2
numba==0.60.0
numpy==2.1.2
Pillow==11.0.0
scikit_learn==1.5.2
scipy==1.14.1
tqdm==4.66.5
